Estimate sample sizes for standalone studies with sensitivity or AUC as endpoint

sampleSize_MRMC This function returns number of cases required for a standalone study for endpoints of sensitivity and AUC.

Usage

sampleSize_Standalone(
  endpoint = "auc",
  theta,
  precision,
  R = 1,
  power = 0.8,
  alpha = 0.05,
  var_auc = "obuchowski",
  corr = FALSE,
  ICC = NULL,
  s = NULL
)

Arguments

endpoint

Character string to inform what is the endpoint of the standalone study. Values can be either auc or sensitivity or specificity. Should be between 0 and 1.

theta

Expected average value of the endpoint. Must be a value between 0 and 1.

precision

Required precision of the point estimate of theta. This is equivalent to half-width of the confidence interval. Must be a numeric value between 0 and 1.

R

Ratio of non-diseased cases to diseased cases. Defaults to 1.

power

Power to detect delta given all other assumptions. Default value is 0.8 corresponding to 80 percent power.

alpha

The type I error rate. Default value is 0.05 corresponding to 5 percent type I error (significance level).

var_auc

Variance estimation method when endpoint is auc. Defaults to the string obuchowski. If value is changed to blume, then method proposed by Blume (2009) will be used to estimate the variance.

corr

Logical value indicating if ICC has to be adjusted (code) or not (FALSE). Defaults to FALSE.

ICC

A numerical value between 0 and 1 indicating the expected ICC if corr is TRUE.

s

Average number of lesions in diseased cases.

Value

A named list

SampleSizeResults - A list containing the sample size results.

References

Flahault A, Cadilhac M, Thomas G. Sample size calculation should be performed for design accuracy in diagnostic test studies. J Clin Epidemiol. 2005 Aug;58(8):859-62. doi: 10.1016/j.jclinepi.2004.12.009. PMID: 16018921.

Zhou, X.-H., Obuchowski, N.A. and McClish, D.K. (2011). Sample Size Calculations. In Statistical Methods in Diagnostic Medicine (eds X.-H. Zhou, N.A. Obuchowski and D.K. McClish). https://doi.org/10.1002/9780470906514.ch6

Author

Dennis Robert dennis.robert.nm@gmail.com

Examples

library("MRMCsamplesize")
result1 <- sampleSize_Standalone(endpoint = "auc", theta = 0.9, precision = 0.05,
                                 R = 1, corr = TRUE, ICC = 0.5, s = 1.25)
#> 
#>      Sample Size Estimation Results 
#> 
#>             ICC = Adjusted for ICC
#>        nUnits_i = 215
#>        nCases_c = 194
#>       nControls = 194
#>          nTotal = 388
#>              DE = 1.125
#>               s = 1.25
#>           power = 0.8
#>           alpha = 0.05
#> 
#> NOTE: 
#> ICC: Is intra-class correlation (ICC) considered while estimating sample size
#> nUnits_i: Number of required units with presence of at least one lesion assuming independence between units
#> nCases_c: Number of required cases with presence of at least one lesion after adjusting for ICC
#> nControls: Number of required controls (images/patients without any lesion)
#> nTotal: Total sample size (cases)
#> DE: Design effect due to ICC
#> s: Assumed average number of lesions in cases with lesions
#> power: Assumed power
#> alpha: Significance level (Type I error rate)
#> 
#> 
result2 <- sampleSize_Standalone(endpoint = "Se", theta = 0.8, precision = 0.05, R = 1)
#> 
#>      Sample Size Estimation Results 
#> 
#>             ICC = No adjustment for ICC
#>        nUnits_i = 503
#>        nCases_c = NA
#>       nControls = 503
#>          nTotal = 1006
#>              DE = NA
#>               s = NA
#>           power = 0.8
#>           alpha = 0.05
#> 
#> NOTE: 
#> ICC: Is intra-class correlation (ICC) considered while estimating sample size
#> nUnits_i: Number of required units with presence of at least one lesion assuming independence between units
#> nCases_c: Number of required cases with presence of at least one lesion after adjusting for ICC
#> nControls: Number of required controls (images/patients without any lesion)
#> nTotal: Total sample size (cases)
#> DE: Design effect due to ICC
#> s: Assumed average number of lesions in cases with lesions
#> power: Assumed power
#> alpha: Significance level (Type I error rate)
#> 
#>